% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_covariates.R
\name{define_covariates}
\alias{define_covariates}
\title{Define Covariates}
\usage{
define_covariates(data_frame, covariates_names, id, cluster = NULL)
}
\arguments{
\item{data_frame}{a data frame that contains the following information}

\item{id}{name of ID column}

\item{cluster}{name of clustering variable}

\item{covariate_names}{a list of the names of columns corresponding to the covariates}

\item{names}{names of the covariates in order of them listed}
}
\value{
a dataframe of covariates linked to numerical ID's
}
\description{
Takes a list of covariates and stores them for future use.
}
\examples{
data(data_example, package = "personlevelEE")

covariates <- define_covariates(covariates =
                                  list(data_example$age, data_example$sex),
                                names = c("age", "sex"), id = data_example$id)

print(covariates)
}
