% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_effect.R
\name{define_effect}
\alias{define_effect}
\title{Define Effect and Incremental Effect Difference}
\usage{
define_effect(data_frame, effect, id, tx, control = 0, treatment = 1)
}
\arguments{
\item{data_frame}{dataframe that contains the information}

\item{effect}{(string) name of effect column}

\item{id}{(string) name of ID column}

\item{tx}{(string) name of treatment column}

\item{control}{Value/Name associated to one treatment (control), default = 0}

\item{treatment}{Value/Name associated to the other treatment (comparator), default = 1}
}
\value{
a list of the regression, ttest objects and associated dataframe
}
\description{
The function will store values for effect, individual ID, and treatment to be used in other functions. Will evaluate the treatment difference
based on values given.
}
\details{
A simple linear regression will evaluate the difference in effect between treatment status.
}
\examples{
data(data_example, package = "personlevelEE")

effect_object <- define_effect(data_example, effect = "effect", id = "id", tx = "tx")

class(effect_object)

print(effect_object)

plot(effect_object, type = "regression")

plot(effect_object, type = "barchart")

plot(effect_object, type = "boxplot")
}
