% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_fieller.R
\name{define_fieller}
\alias{define_fieller}
\title{Run Fieller's theorem}
\usage{
define_fieller(cost, effect)
}
\arguments{
\item{cost}{a [define_cost()] object}

\item{effect}{a [define_effect()] object}
}
\value{
a data frame containing the data to create confidence ellipses
}
\description{
Function uses Fieller's theorem to define values to create a confidence intervals and related confidence ellipses.
}
