% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personlevelEE.R
\docType{package}
\name{personlevelEE}
\alias{personlevelEE}
\title{Person Level Health Economic Modelling}
\description{
For the use of economic evaluations using person - level data
(clinical RCTs, health administrative, etc). Includes cost effectiveness
regression modelling, net benefit approaches, and both parametric and non parametric uncertainty analyses.
}
