% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cost.R
\name{plot.define_cost}
\alias{plot.define_cost}
\title{Plot Define Cost Results}
\usage{
\method{plot}{define_cost}(object, type = c("regression", "barchart", "boxplot"))
}
\arguments{
\item{object}{A result of [define_cost()]}

\item{type}{Type of plot}
}
\value{
plots
}
\description{
Plot Define Cost Results
}
