% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_fieller.R
\name{plot.define_fieller}
\alias{plot.define_fieller}
\title{Plot Confidence Ellipses}
\usage{
\method{plot}{define_fieller}(x, ...)
}
\arguments{
\item{x}{a [define_fieller()] object}

\item{...}{additional arguments affecting the plot
produced.}
}
\value{
a [ggplot2()] object
}
\description{
Plot Confidence Ellipses
}
