% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bootstrap.R
\name{plot.run_bootstrap}
\alias{plot.run_bootstrap}
\title{Plot output of bootstrap}
\usage{
\method{plot}{run_bootstrap}(x, type = c("cloud", "ceac"), ...)
}
\arguments{
\item{x}{object of [run_bootstrap()]}

\item{type}{type of graph to show, either "cloud" or "ceac"}

\item{...}{additional arguments affecting the plot
produced.}
}
\value{
a [ggplot2()] object
}
\description{
Plot output of bootstrap
}
