% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_INB_model.R
\name{plot.run_INB_model}
\alias{plot.run_INB_model}
\title{Plot INB model}
\usage{
\method{plot}{run_INB_model}(object, type = c("regression", "barchart", "boxplot"))
}
\arguments{
\item{object}{a [run_INB_model()] object}

\item{type}{type of graph to plot, see more details}
}
\description{
Plot INB model
}
