% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{plot_regression}
\alias{plot_regression}
\title{plot regression model 1}
\usage{
plot_regression(model)
}
\arguments{
\item{model}{regression model}
}
\description{
plot regression model 1
}
\keyword{internal}
