% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plots.R
\name{plot_regression_bw}
\alias{plot_regression_bw}
\title{plot regression model 2}
\usage{
plot_regression_bw(model)
}
\arguments{
\item{model}{regression model}
}
\description{
plot regression model 2
}
\keyword{internal}
