% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_covariates.R
\name{print.define_covariates}
\alias{print.define_covariates}
\title{Print Covariates}
\usage{
\method{print}{define_covariates}(x, ...)
}
\arguments{
\item{x}{an [define_covariates()] object}

\item{...}{}
}
\description{
Print Covariates
}
