% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cost.R, R/define_effect.R
\name{print.define_effect}
\alias{print.define_effect}
\title{Summarize Define Cost Results}
\usage{
\method{print}{define_effect}(object)

\method{print}{define_effect}(object)
}
\arguments{
\item{object}{a [define_effect()] object}
}
\description{
Printing a formatted version of the regression results in context of an incremental cost effectiveness ratio for economic evaluations.

Printing a formatted version of the regression results in context of an incremental cost effectiveness ratio for economic evaluations.
}
\details{
To view raw regression results, see [summary()]

To view raw regression results, see [summary()]
}
