% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_fieller.R
\name{print.define_fieller}
\alias{print.define_fieller}
\title{Print NB data}
\usage{
\method{print}{define_fieller}(x, ...)
}
\arguments{
\item{x}{an \code{\link[=define_fieller]{define_fieller()}} object}

\item{...}{additional arguments affecting the print}
}
\description{
Print NB data
}
