% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icer_model.R
\name{print.run_icer_model}
\alias{print.run_icer_model}
\title{Print ICER regression results}
\usage{
\method{print}{run_icer_model}(object)
}
\arguments{
\item{object}{an [run_icer_model()] object}
}
\value{
returns string
}
\description{
Print ICER regression results
}
