% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_INB_model.R
\name{print.run_INB_model}
\alias{print.run_INB_model}
\title{Print INB regression results}
\usage{
\method{print}{run_INB_model}(x, ...)
}
\arguments{
\item{x}{an [run_INB_model()] object}

\item{...}{additional arguments affecting the summary
produced.}
}
\value{
returns formatted regression results
}
\description{
Print INB regression results
}
