% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_pub_bw}
\alias{theme_pub_bw}
\title{theme ggplot 1}
\usage{
theme_pub_bw(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{font family}
}
\description{
theme ggplot 1
}
\keyword{internal}
