% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_pub_bw1}
\alias{theme_pub_bw1}
\title{theme ggplot 2}
\usage{
theme_pub_bw1(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{font family}
}
\description{
theme ggplot 2
}
\keyword{internal}
