% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cost.R
\name{define_cost}
\alias{define_cost}
\title{Define Cost Values]}
\usage{
define_cost(cost, id, tx)
}
\arguments{
\item{cost}{Vector of numerical values referring to cost}

\item{id}{Vector of numerical values referring to individual ID}

\item{tx}{}
}
\value{
A data frame of a cost values associated to an individual ID
}
\description{
Define the vector used for the values to measure effect.
}
\details{
Function will
}
