% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_covariates.R
\name{define_covariates}
\alias{define_covariates}
\title{Define Covariates}
\usage{
define_covariates(data_frame, covariates_names, id, cluster = "none")
}
\arguments{
\item{data_frame}{a data frame that contains the following information}

\item{id}{(string) name of ID column}

\item{cluster}{name of clustering variable}

\item{covariate_names}{a list of the names of columns corresponding to the covariates}
}
\value{
a dataframe of covariates linked to numerical ID's
}
\description{
Takes a list of covariates and stores them for future use.
}
\examples{
data(data_example, package = "personlevelEE")

covariates <- define_covariates(data_frame = data_example, covariates_names = c("age", "sex"), id = "id")

print(covariates)
}
