% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_covariates.R
\name{define_covariates}
\alias{define_covariates}
\title{Define Covariates}
\usage{
define_covariates(covariates, names, id)
}
\arguments{
\item{covariates}{a list of vectors corresponding to the covariates}

\item{names}{list of names of the covariates}

\item{id}{vector of numerical ID's}
}
\value{
a dataframe of covariates linked to numerical ID's
}
\description{
Takes a list of covariates and stores them for future use.
}
