% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_effect.R
\name{define_effect}
\alias{define_effect}
\title{Define Effect Values}
\usage{
define_effect(effect, id, tx)
}
\arguments{
\item{tx}{}
}
\value{
A data.frame
}
\description{
Define the vector for the values used to measure the effect.
}
