% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_fieller.R
\name{define_fieller}
\alias{define_fieller}
\title{Run Fieller's theorem}
\usage{
define_fieller(cost, effect)
}
\arguments{
\item{cost}{a \code{\link[=define_cost]{define_cost()}} object}

\item{effect}{a \code{\link[=define_effect]{define_effect()}} object}
}
\value{
a data frame containing the data to create confidence ellipses
}
\description{
Function uses Fieller's theorem to define values to create a confidence intervals and related confidence ellipses.
}
\examples{
data(data_example, package = "personlevelEE")

cost_object <- define_cost(data_example, cost = "cost", id = "id", tx = "tx")

effect_object <- define_effect(data_example, effect = "effect", id = "id", tx = "tx")

fieller <- define_fieller(cost = cost_object, effect = effect_object)

plot(fieller)

plot(fieller, CI = TRUE, bw = TRUE)
print(fieller) ## 95\% confidence interval
}
