% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_NB.R
\name{define_NB}
\alias{define_NB}
\title{Define net benefit (NB) values}
\usage{
define_NB(lambda, cost, effect)
}
\arguments{
\item{lambda}{a value for WTP threshold}

\item{cost}{a [define_cost()] object}

\item{effect}{a [define_effect()] object}
}
\value{
return dataframe that includes NB values associated with an individual id based on given lambda
}
\description{
Creates a smaller
}
