% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_effect.R
\name{plot.define_effect}
\alias{plot.define_effect}
\title{Plot Define Effect Results}
\usage{
\method{plot}{define_effect}(x, type = c("regression", "barchart", "boxplot"), bw = FALSE, ...)
}
\arguments{
\item{x}{A result of \code{\link[=define_effect]{define_effect()}}}

\item{type}{Type of plot}

\item{bw}{Black & white plot theme for publications}
}
\value{
a \code{\link[=ggplot2]{ggplot2()}} object
}
\description{
Plot Define Effect Results
}
\examples{
data(data_example, package = "personlevelEE")

effect_object <- define_effect(data_example$effect, data_example$id, data_example$tx)

class(effect_object)

print(effect_object)

plot(effect_object, type = "regression")

plot(effect_object, type = "barchart")

plot(effect_object, type = "boxplot")
}
