% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_effect.R
\name{plot.define_effect}
\alias{plot.define_effect}
\title{Plot Define Effect Results}
\usage{
\method{plot}{define_effect}(object, type = c("regression", "barchart", "boxplot"))
}
\arguments{
\item{object}{A result of [define_effect()]}

\item{type}{Type of plot}
}
\value{
plots
}
\description{
Plot Define Effect Results
}
