% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bootstrap.R
\name{plot.run_bootstrap}
\alias{plot.run_bootstrap}
\title{Plot output of bootstrap}
\usage{
\method{plot}{run_bootstrap}(object, type = c("cloud", "ceac"))
}
\arguments{
\item{type}{}
}
\description{
Plot output of bootstrap
}
