% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icer_model.R
\name{plot.run_icer_model}
\alias{plot.run_icer_model}
\title{Plot ICER regressions}
\usage{
\method{plot}{run_icer_model}(object, type = c("regression"))
}
\arguments{
\item{object}{an [run_icer_model()] object}

\item{type}{type of graph, default = regression diagnostics}
}
\value{
a [ggplot()] object
}
\description{
Plot ICER regressions
}
