% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_INB_model.R
\name{plot.run_inb_model}
\alias{plot.run_inb_model}
\title{Plot INB model}
\usage{
\method{plot}{run_inb_model}(x, type = c("regression", "barchart", "boxplot", "ceac"), bw = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link[=run_INB_model]{run_INB_model()}} object}

\item{type}{type of graph to plot ("regression", "barchart", "boxplot", "ceac")}

\item{bw}{Black & white plot theme for publications}

\item{...}{additional arguments affecting the summary
produced.}
}
\value{
a \code{\link[=ggplot2]{ggplot2()}} object
}
\description{
Plot INB model
}
\examples{
library(tidyverse)
data(data_example, package = "personlevelEE")

cost_object <- define_cost(data_example, cost = cost, id = id, tx = tx)

effect_object <- define_effect(data_example, effect = effect, id = id, tx = tx)

##add cluster variable
data_example <- data_example \%>\% mutate(cluster_v = sample(1:4, length(data_example$tx), replace = TRUE))
covariates <- define_covariates(data_frame = data_example, covariates_names = c("age", "sex"), id = id, cluster ="cluster_v")


nb_value <- define_NB(lambdas = c(100, 350, 1000, 10000),
                      cost = cost_object, effect = effect_object)

nb_model <- run_INB_model(nb_value, covariates = covariates)

print(nb_model)

plot(nb_model, type = "regression", bw = TRUE)

plot(nb_model, type = "boxplot", bw = TRUE)

plot(nb_model, type = "barchart", bw = TRUE)

plot(nb_model, type = "ceac", bw = TRUE)




}
