% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_INB_model.R
\name{plot.run_INB_model}
\alias{plot.run_INB_model}
\title{Plot INB model}
\usage{
\method{plot}{run_INB_model}(x, ..., type = c("regression", "barchart", "boxplot"))
}
\arguments{
\item{x}{a [run_INB_model()] object}

\item{...}{additional arguments affecting the summary
produced.}

\item{type}{type of graph to plot, see more details}
}
\value{
a [ggplot2()] object
}
\description{
Plot INB model
}
