% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cost.R
\name{print.define_cost}
\alias{print.define_cost}
\title{Summarize Define Cost Results}
\usage{
\method{print}{define_cost}(x, ...)
}
\arguments{
\item{x}{a [define_cost()] object}

\item{...}{additional arguments affecting the summary
produced.}
}
\description{
Printing a formatted version of the regression results in context of an incremental cost effectiveness ratio for economic evaluations.
}
\details{
To view raw regression results, see [summary()]
}
