% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_effect.R
\name{print.define_effect}
\alias{print.define_effect}
\title{Summarize Define Effect Results}
\usage{
\method{print}{define_effect}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=define_effect]{define_effect()}} object}

\item{...}{additional arguments affecting the summary
produced.}
}
\value{
formatted regression results
}
\description{
Printing a formatted version of the regression results in context of an incremental cost effectiveness ratio for economic evaluations.
}
\details{
To view raw regression results, see \code{\link[=summary]{summary()}}
}
