% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_NB.R
\name{print.define_NB}
\alias{print.define_NB}
\title{Print NB data}
\usage{
\method{print}{define_NB}(x, ...)
}
\arguments{
\item{x}{an \code{\link[=define_NB]{define_NB()}} object}

\item{...}{additional arguments affecting the print.}
}
\description{
Print NB data
}
