% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icer_model.R
\name{print.run_icer_model}
\alias{print.run_icer_model}
\title{Print ICER regression results}
\usage{
\method{print}{run_icer_model}(x, ...)
}
\arguments{
\item{x}{an \code{\link[=run_icer_model]{run_icer_model()}} object}

\item{...}{additional arguments affecting the print
produced.}
}
\value{
returns formatted table of regression results
}
\description{
Print ICER regression results
}
