% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bootstrap.R
\name{run_bootstrap}
\alias{run_bootstrap}
\title{Run bootstrap}
\usage{
run_bootstrap(n, cost, effect)
}
\arguments{
\item{n}{number of repetitions for bootstrapping}

\item{cost}{a [define_cost()] object}

\item{effect}{a [define_effect()] object}
}
\value{
a [dataframe()] of bootstrapped results
}
\description{
Non-parametric uncertainty analysis: bootstrapping. This function will sample the ICER estimate n times.
}
