% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icer_model.R
\name{run_icer_model}
\alias{run_icer_model}
\title{Run ICER model}
\usage{
run_icer_model(
  cost,
  effect,
  covariates = NULL,
  interaction = FALSE,
  method = "linear"
)
}
\arguments{
\item{cost}{result of \code{\link[=define_cost]{define_cost()}}}

\item{effect}{result of \code{\link[=define_effect]{define_effect()}}}

\item{covariates}{a \code{\link[=define_covariates]{define_covariates()}} object}

\item{interaction}{boolean, if model should consider interaction terms, default = FALSE}

\item{method}{type of model ("linear", "hierarchical"), default = "linear"}
}
\value{
a list containing the regression results for cost, effect, related data frames, and final icer values
}
\description{
This function will calculate an ICER based on the cost and effect regressions.
It will consider the different covariates and interaction terms from your data in the model.
}
\examples{
data(data_example, package = "personlevelEE")

cost_object <- define_cost(data_example, cost = "cost", id = "id", tx = "tx")

effect_object <- define_effect(data_example, effect = "effect", id = "id", tx = "tx")

covariates <- define_covariates(data_frame = data_example, covariates_names = c("age", "sex"), id = id)

icer_model <- run_icer_model(cost_object, effect_object,
                             covariates = covariates, interaction = FALSE)

print(icer_model)

summary(icer_model)

plot(icer_model, type = "regression")
plot(icer_model, type = "regression", bw = TRUE)
}
