% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icer_model.R
\name{run_icer_model}
\alias{run_icer_model}
\title{Run ICER model}
\usage{
run_icer_model(cost, effect, covariates, clustering, method, interaction)
}
\arguments{
\item{interaction}{}
}
\description{
Run ICER model
}
