% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icer_model.R
\name{run_icer_model}
\alias{run_icer_model}
\title{Run ICER model}
\usage{
run_icer_model(
  cost,
  effect,
  covariates = NA,
  interaction = TRUE,
  method = "linear"
)
}
\arguments{
\item{cost}{result of [define_cost()]}

\item{effect}{result of [define_effect()]}

\item{covariates}{a dataframe of covariates as defined by [define_covariates()]}

\item{interaction}{TRUE, if model should consider interaction terms}

\item{method}{type of model ("linear", "hierarchical")}

\item{clustering}{clustering variable, default = NA}
}
\description{
This function will
}
