% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_INB_model.R
\name{run_INB_model}
\alias{run_INB_model}
\title{Run Incremental Net Benefit Model}
\usage{
run_INB_model(nb_value, covariates = NA)
}
\arguments{
\item{nb_value}{a [define_nb()] object}

\item{covariates}{a [define_covariates()] object}
}
\description{
Run Incremental Net Benefit Model
}
