% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icer_model.R
\name{summary.run_icer_model}
\alias{summary.run_icer_model}
\title{Print a summary table of ICER}
\usage{
\method{summary}{run_icer_model}(x, ...)
}
\arguments{
\item{x}{an [run_icer_model()] object}

\item{...}{additional arguments affecting the summary
produced.}
}
\value{
a tibble
}
\description{
Print a summary table of ICER
}
